requirements.txt
pyyaml
yfinance
requests